requirements.txt
keras==2.2.4
tensorflow-gpu==2.0.0a0
numpy
opencv-python
matplotlib
scikit-image